requirements.txt
streamlit
tensorflow==2.16.1
numpy
Pillow
opencv-python
h5py
